further clean ups
